refactor: use types instead of is
